refactor date range to exclude end date in ActivityReportService
