+++ src/main/java/depromeet/lessonfour/server/activityrecord/infra/repository/QueryDslActivityRecordRepository.java
@@ -63,7 +63,7 @@ public class QueryDslActivityRecordRepository {
         .where(
             activityRecord.userId.eq(userId),
             activityRecord.activityAt.date.goe(start.toDate()),
-            activityRecord.activityAt.date.loe(end.toDate()),
+            activityRecord.activityAt.date.lt(end.toDate()),
             activityRecord.isDeleted.eq(false))
         .fetch();
   }
