+++ src/main/java/depromeet/lessonfour/server/report/app/client/ActivityRecordClient.java
@@ -1,4 +1,4 @@
-package depromeet.lessonfour.server.report.app.service;
+package depromeet.lessonfour.server.report.app.client;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import depromeet.lessonfour.server.activityrecord.domain.entity.ActivityRecord;
 
-public interface ActivityRecordService {
+public interface ActivityRecordClient {
 
   Map<LocalDate, ActivityRecord> getActivityRecordsBetween(
       Long userId, LocalDateTime start, LocalDateTime end);
