+++ src/main/java/depromeet/lessonfour/server/report/app/service/ActivityReportService.java
@@ -2,12 +2,12 @@ package depromeet.lessonfour.server.report.app.service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.Map;
 
 import org.springframework.stereotype.Service;
 
 import depromeet.lessonfour.server.activityrecord.domain.entity.ActivityRecord;
+import depromeet.lessonfour.server.report.app.client.ActivityRecordClient;
 import depromeet.lessonfour.server.report.domain.service.ActivityEvaluationService;
 import depromeet.lessonfour.server.report.domain.vo.ActivityReport;
 import lombok.RequiredArgsConstructor;
@@ -16,14 +16,14 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ActivityReportService {
 
-  private final ActivityRecordService activityRecordService;
+  private final ActivityRecordClient activityRecordClient;
   private final ActivityEvaluationService activityEvaluationService;
 
   public ActivityReport generateDailyReport(Long userId, LocalDateTime baseDatetime) {
     LocalDate baseDate = baseDatetime.toLocalDate();
     Map<LocalDate, ActivityRecord> recordsByDate =
-        activityRecordService.getActivityRecordsBetween(
-            userId, baseDate.minusDays(1).atStartOfDay(), baseDate.atTime(LocalTime.MAX));
+        activityRecordClient.getActivityRecordsBetween(
+            userId, baseDate.minusDays(1).atStartOfDay(), baseDate.plusDays(1).atStartOfDay());
 
     ActivityRecord currentRecord = recordsByDate.get(baseDate);
     ActivityRecord previousRecord = recordsByDate.get(baseDate.minusDays(1));
